Added support for Gradle multi-level projects
